+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# dev
+- fix intrinsic metric extension in MPLA0
+
 # 0.3.1
 - fix missing unix-style arguments