+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # dev
 - fix intrinsic metric extension in MPLA0
+- fix default metric weights config
 
 # 0.3.1
 - fix missing unix-style arguments